# openapi.yaml
openapi: 3.0.3
info:
  title: Parkmanshift API
  version: 1.0.0
  description: Spécification de l'API pour le backend Parkmanshift — contrat d'interface partagé backend/web.
servers:
  - url: http://localhost:8080
    description: Serveur local par défaut
tags:
  - name: hello
    description: Walking skeleton endpoint
  - name: auth
    description: Authentification et gestion des utilisateurs
  - name: parking
    description: Gestion du parking
security:
  - BearerAuth: []
paths:
  /api/hello:
    get:
      security: []
      tags:
        - hello
      summary: Récupérer un message de salutation (walking skeleton)
      operationId: hello
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: "Walking Skeleton"
  /api/auth/login:
    post:
      security: []
      tags:
        - auth
      summary: Se connecter pour obtenir un token JWT
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
  /api/auth/register-employee:
    post:
      tags:
        - auth
      summary: Enregistrer un nouvel employé (réservé aux secrétaires)
      operationId: registerEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationDto'
      responses:
        '200':
          description: Utilisateur créé avec succès
  /api/users/search:
    get:
      tags:
        - auth
      summary: Rechercher des utilisateurs (réservé aux secrétaires)
      operationId: searchUsers
      parameters:
        - name: query
          in: query
          description: Terme de recherche (nom d'utilisateur)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
  /api/users/change-password:
    post:
      tags:
        - auth
      summary: Changer son mot de passe
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Mot de passe changé avec succès
        '400':
          description: Mot de passe actuel invalide
  /api/parking/state:
    get:
      tags:
        - parking
      summary: Récupérer l'état des places de parking
      operationId: getParkingState
      parameters:
        - name: date
          in: query
          description: Date pour laquelle récupérer l'état (par défaut la date du jour si omis)
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpotStateDto'
  /api/parking/reservations:
    post:
      tags:
        - parking
      summary: Réserver une place de parking
      operationId: reserveSpot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReserveRequestDto'
      responses:
        '200':
          description: Réservation créée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
  /api/parking/reservations/history:
    get:
      tags:
        - parking
      summary: Récupérer l'historique des réservations d'un employé
      description: Les employés ne peuvent consulter que leur propre historique. Les secrétaires et managers peuvent consulter n'importe quel historique en passant l'ID de l'employé.
      operationId: getReservationHistory
      parameters:
        - name: employeeId
          in: query
          description: ID de l'employé dont on veut l'historique (optionnel pour l'utilisateur courant)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationDto'
  /api/parking/reservations/{id}/checkin:
    post:
      tags:
        - parking
      summary: Signaler l'arrivée (check-in) pour une réservation
      operationId: checkIn
      parameters:
        - name: id
          in: path
          required: true
          description: ID (UUID) de la réservation
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Check-in effectué avec succès
  /api/parking/reservations/{id}:
    delete:
      tags:
        - parking
      summary: Annuler une réservation
      operationId: cancelReservation
      parameters:
        - name: id
          in: path
          required: true
          description: ID (UUID) de la réservation à annuler
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Annulation effectuée avec succès
    put:
      tags:
        - parking
      summary: Modifier une réservation
      operationId: updateReservation
      parameters:
        - name: id
          in: path
          required: true
          description: ID (UUID) de la réservation à modifier
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReserveRequestDto'
      responses:
        '200':
          description: Modification effectuée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
  /api/parking/public/check-in/verify:
    get:
      security: []
      tags:
        - parking
      summary: Vérifier un code de check-in pour une place
      operationId: verifyCheckIn
      parameters:
        - name: spotLabel
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckInVerificationDto'
  /api/parking/public/check-in/confirm:
    post:
      security: []
      tags:
        - parking
      summary: Confirmer le check-in avec le code
      operationId: confirmCheckIn
      parameters:
        - name: spotLabel
          in: query
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Check-in confirmé

  /api/parking/dashboard:
    get:
      tags:
        - parking
      summary: Récupérer les statistiques du dashboard manager
      operationId: getDashboardStats
      parameters:
        - name: yearMonth
          in: query
          description: Mois pour filtrer les statistiques (ex. 2026-04)
          required: false
          schema:
            type: string
        - name: employeeId
          in: query
          description: ID de l'employé pour filtrer les statistiques
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStatsDto'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SpotStateDto:
      type: object
      properties:
        spot:
          $ref: '#/components/schemas/ParkingSpotDto'
        status:
          $ref: '#/components/schemas/SpotStatusDto'
        reservedBy:
          type: string
          description: Username of the person who reserved/occupies this spot (null if free)
        reservationStatus:
          $ref: '#/components/schemas/ReservationStatusDto'
        reservationId:
          type: string
          format: uuid
          description: ID of the reservation (null if free)
    ParkingSpotDto:
      type: object
      properties:
        label:
          type: string
          example: "P-01"
        type:
          $ref: '#/components/schemas/SpotTypeDto'
    ReservationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        parkingSpotLabel:
          type: string
          example: "P-01"
        employeeId:
          type: string
          example: "EMP-1234"
        date:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/ReservationStatusDto'
    ReserveRequestDto:
      type: object
      required:
        - parkingSpotLabel
        - date
      properties:
        parkingSpotLabel:
          type: string
          example: "P-01"
        date:
          type: string
          format: date
        targetUsername:
          type: string
          description: Réserver pour un autre employé (réservé aux secrétaires)
    UserRegistrationDto:
      type: object
      required:
        - username
        - fullName
        - password
      properties:
        username:
          type: string
        fullName:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [EMPLOYEE, MANAGER, SECRETARY]
          default: EMPLOYEE
    SpotStatusDto:
      type: string
      enum:
        - FREE
        - RESERVED
        - OCCUPIED
        - UNAVAILABLE
    SpotTypeDto:
      type: string
      enum:
        - ELECTRIC
        - THERMAL
    ReservationStatusDto:
      type: string
      enum:
        - RESERVED
        - OCCUPIED
        - CANCELLED
    DashboardStatsDto:
      type: object
      properties:
        occupancyRate:
          type: number
          format: double
          description: Taux d'occupation moyen ou actuel (%)
        noShowProportion:
          type: number
          format: double
          description: Proportion de réservations non utilisées (%)
        electricSpotProportion:
          type: number
          format: double
          description: Proportion de places électriques utilisées (%)
        totalReservations:
          type: integer
          format: int32
          description: Nombre total de réservations sur la période
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [EMPLOYEE, MANAGER, SECRETARY]
        checkInCode:
          type: string
          pattern: '^[0-9]{4}$'
    CheckInVerificationDto:
      type: object
      properties:
        username:
          type: string
        fullName:
          type: string
        existingReservationSpotLabel:
          type: string
          nullable: true
        conflict:
          type: boolean